# api-spec.json
{
  "components": {
    "schemas": {
...
      "ImageThumbnailSchemaOut": {
        "properties": {
          "id": {
            "description": "The unique identifier of the object",
            "title": "Id",
            "type": "integer"
          },
          "thumbnail_height": {
            "title": "Thumbnail Height",
            "type": "integer"
          },
          "thumbnail_url": {
            "title": "Thumbnail Url",
            "type": "string"
          },
          "thumbnail_width": {
            "title": "Thumbnail Width",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["id", "title", "thumbnail_url", "thumbnail_height", "thumbnail_width"],
        "title": "ImageThumbnailSchemaOut",
        "type": "object"
      },
...
      "PagedImageThumbnailSchemaOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImageThumbnailSchemaOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": ["items", "count"],
        "title": "PagedImageThumbnailSchemaOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AsyncSessionAuthIsActive": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      },
      "AsyncSessionAuthIsActiveSuperUserOrStaff": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      },
      "SessionAuthIsActive": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      },
      "SessionAuthIsActiveSuperUserOrStaff": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "",
    "title": "Memoria API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/image/": {
      "get": {
        "operationId": "list_images",
        "parameters": [
          {
            "description": "Filter by dirty status",
            "in": "query",
            "name": "is_dirty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dirty status",
              "title": "Is Dirty"
            }
          },
          {
            "description": "Filter by starred status",
            "in": "query",
            "name": "is_starred",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by starred status",
              "title": "Is Starred"
            }
          },
          {
            "description": "Filter by deletion status",
            "in": "query",
            "name": "is_deleted",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by deletion status",
              "title": "Is Deleted"
            }
          },
          {
            "description": "Filter by ImageFolder ID",
            "in": "query",
            "name": "folder_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ImageFolder ID",
              "title": "Folder Id"
            }
          },
          {
            "description": "Filter by Person IDs",
            "in": "query",
            "name": "people_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Person IDs",
              "title": "People Ids"
            }
          },
          {
            "description": "Filter by Pet IDs",
            "in": "query",
            "name": "pets_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Pet IDs",
              "title": "Pets Ids"
            }
          },
          {
            "description": "Filter by Tag IDs",
            "in": "query",
            "name": "tags_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Tag IDs",
              "title": "Tags Ids"
            }
          },
          {
            "description": "Filter images from this date onwards using comparison_date",
            "in": "query",
            "name": "date_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this date onwards using comparison_date",
              "title": "Date Start"
            }
          },
          {
            "description": "Filter images up to this date using comparison_date",
            "in": "query",
            "name": "date_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images up to this date using comparison_date",
              "title": "Date End"
            }
          },
          {
            "description": "Filter images from this year onwards",
            "in": "query",
            "name": "year_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this year onwards",
              "title": "Year Start"
            }
          },
          {
            "description": "Filter images up to this year",
            "in": "query",
            "name": "year_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images up to this year",
              "title": "Year End"
            }
          },
          {
            "description": "Filter images from this month onwards (1-12, includes null months)",
            "in": "query",
            "name": "month_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this month onwards (1-12, includes null months)",
              "title": "Month Start"
            }
          },
          {
            "description": "Filter images up to this month (1-12, includes null months)",
            "in": "query",
            "name": "month_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images up to this month (1-12, includes null months)",
              "title": "Month End"
            }
          },
          {
            "description": "Filter images from this day onwards (1-31, includes null days)",
            "in": "query",
            "name": "day_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this day onwards (1-31, includes null days)",
              "title": "Day Start"
            }
          },
          {
            "description": "Filter images up to this day (1-31, includes null days)",
            "in": "query",
            "name": "day_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images up to this day (1-31, includes null days)",
              "title": "Day End"
            }
          },
          {
            "description": "Filter images from this exact year",
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this exact year",
              "title": "Year"
            }
          },
          {
            "description": "Filter images from this exact month (1-12)",
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this exact month (1-12)",
              "title": "Month"
            }
          },
          {
            "description": "Filter images from this exact day (1-31)",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 31,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images from this exact day (1-31)",
              "title": "Day"
            }
          },
          {
            "description": "Filter by country code (ISO 3166-1 alpha 2)",
            "in": "query",
            "name": "country_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by country code (ISO 3166-1 alpha 2)",
              "title": "Country Code"
            }
          },
          {
            "description": "Filter by state/province code (ISO 3166-2)",
            "in": "query",
            "name": "subdivision_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by state/province code (ISO 3166-2)",
              "title": "Subdivision Code"
            }
          },
          {
            "description": "Filter by city name (partial match)",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by city name (partial match)",
              "title": "City"
            }
          },
          {
            "description": "Filter by sub-location (partial match)",
            "in": "query",
            "name": "sub_location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by sub-location (partial match)",
              "title": "Sub Location"
            }
          },
          {
            "description": "Field to sort by",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "pk",
              "description": "Field to sort by",
              "enum": ["created", "-created", "modified", "-modified", "pk", "title", "-title"],
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedImageThumbnailSchemaOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "List Images",
        "tags": ["images"]
      }
    },
    "/api/image/bulk/thumbnails/": {
      "post": {
        "operationId": "image_get_thumbnails_bulk_info",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "title": "Image Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageThumbnailSchemaOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image Thumbnails Bulk Info",
        "tags": ["images"]
      }
    },
    "/api/image/{image_id}/date/": {
      "get": {
        "operationId": "image_get_date",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDateSchemaOut"
                }
              }
            },
            "description": "OK"
          },
          "204": {
            "description": "The image has no date"
          },
          "401": {
            "description": "The user does not have permissions for this image"
          },
          "404": {
            "description": "The image does not exist"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image Date",
        "tags": ["images"]
      },
      "patch": {
        "operationId": "image_update_date",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageDateUpdateSchemaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDateSchemaOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "The user does not have permissions for this image"
          },
          "404": {
            "description": "The image does not exist"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Update Image Date",
        "tags": ["images"]
      }
    },
    "/api/image/{image_id}/location/": {
      "get": {
        "operationId": "image_get_location",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageLocationSchemaOut"
                }
              }
            },
            "description": "OK"
          },
          "204": {
            "description": "The image has no location"
          },
          "401": {
            "description": "The user does not have permissions for this image"
          },
          "404": {
            "description": "The image does not exist"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image Location",
        "tags": ["images"]
      },
      "patch": {
        "operationId": "image_update_location",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageLocationUpdateSchemaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageLocationSchemaOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The country is not found or subdivision is not in the country"
          },
          "401": {
            "description": "The user does not have permissions for this image"
          },
          "404": {
            "description": "The image does not exist"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Update Image Location",
        "tags": ["images"]
      }
    },
    "/api/image/{image_id}/metadata/": {
      "get": {
        "operationId": "image_get_metadata",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadataSchemaOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image Details",
        "tags": ["images"]
      },
      "patch": {
        "operationId": "image_update_metadata",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMetadataUpdateSchemaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadataSchemaOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Update Image Details",
        "tags": ["images"]
      }
    },
    "/api/image/{image_id}/people/": {
      "get": {
        "operationId": "image_get_people",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonInImageSchemaOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image People",
        "tags": ["images"]
      }
    },
    "/api/image/{image_id}/pets/": {
      "get": {
        "operationId": "image_get_pets",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PetInImageSchemaOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuthIsActive": []
          }
        ],
        "summary": "Get Image Pets",
        "tags": ["images"]
      }
    },
...
  },
  "servers": []
}